default g:zv_docsets_dir; rename g:zv_zeal_directory
This patch adds default values for Zeal's docsets directory location and
renames the `g:zv_zeal_directory` variable to `g:zv_zeal_executable`
because its value is supposed to specify the path of the Zeal
executable, not the directory which contains the Zeal executable.

This patch also simplifies the implementation of GetDocsetNameFromDir(),
GetDocsetName(), and ExecuteZeal() functions.  Notably, I used system()
with shellescape() instead of `:!` to avoid the `:redraw!` afterwards.